Write playRound() and edit getHumanChoice()
Create function that calls getComputerChoice() and getHumanChoice(), and
stores the return values into local humanChoice and computerChoice variables.

Then, use switch statement to compare the variables, determine a winner,
console.log with message of outcome, increment the winner's score, and
console.log the score.

Also edited getHumanChoice() to return the choice value instead of
console.log it, and to call itself again when given invalid input.