Restructure playRound()
Decrease amount of cases for switch statement in playRound() by
matching humanChoice instead of true, then adding if/else statements
for comparison with computerChoice.

Also, change default for getHumanChoice() to throw an Error for
invalid input - for simplicity, but also because recursively
calling the function caused the value stored in humanChoice to
be undefined.